the DSB normalization function

the DSB normalization function

DSBNormalizeProtein(
  cell_protein_matrix,
  empty_drop_matrix,
  define.pseudocount = FALSE,
  pseudocount.use,
  denoise.counts = TRUE,
  use.isotype.control = FALSE,
  isotype.control.name.vec = NULL
)

Arguments

cell_protein_matrix
the data to be normalized cells = columns, proteins = rows

empty_drop_matrix
the empty droplets used for background correction

define.pseudocount
TRUE FALSE : false by default users can supply a pseudocount besides the default 10 which is optimal for CITEseq data.

pseudocount.use
the pseudocount to use if overriding the default pseudocount by setting efine.pseudocount to TRUE

denoise.counts
account for a per cell background correction factor by getting the mean of negative proteins in that cell + or - isotype controls

use.isotype.control
should the denoising covariate include isotype controls - recommended but false by default

isotype.control.name.vec
a vector of the names of the isotype control proteins exactly as in the data to be normalized

Value

a normalized matrix that can be added to any S4 object e.g. Seurat or SingleCellExperiment Object